+++ packages/miniflare/src/plugins/core/index.ts
@@ -216,6 +216,10 @@ export const CoreSharedOptionsSchema = z.object({
 
 	liveReload: z.boolean().optional(),
 
+	// Enable auto service / durable objects discovery with the dev registry
+	unsafeDevRegistryPath: z.string().optional(),
+	// Enable External Durable Objects Proxy / Internal DOs registration
+	unsafeDevRegistryDurableObjectProxy: z.boolean().default(false),
 	// This is a shared secret between a proxy server and miniflare that can be
 	// passed in a header to prove that the request came from the proxy and not
 	// some malicious attacker.