+++ packages/wrangler/src/api/startDevWorker/LocalRuntimeController.ts
@@ -127,6 +127,7 @@ export async function convertToConfigBundle(
 		imagesLocalMode: event.config.dev?.imagesLocalMode ?? false,
 		testScheduled: !!event.config.dev.testScheduled,
 		tails: event.config.tailConsumers,
+		containers: event.config.containers ?? {},
 	};
 }
 
@@ -171,7 +172,8 @@ export class LocalRuntimeController extends RuntimeController {
 					this.#log,
 					configBundle,
 					this.#proxyToUserWorkerAuthenticationSecret,
-					this.#mixedModeSession?.mixedModeConnectionString
+					this.#mixedModeSession?.mixedModeConnectionString,
+					!!getFlag("MIXED_MODE")
 				);
 			options.liveReload = false; // TODO: set in buildMiniflareOptions once old code path is removed
 			if (this.#mf === undefined) {