+++ packages/wrangler/src/config/environment.ts
@@ -103,6 +103,11 @@ export type ContainerApp = {
 	 *  - manual: The container application will be rollout fully by manually actioning progress steps.
 	 */
 	rollout_kind?: "full_auto" | "none" | "full_manual";
+
+	/**
+	 * Ports to be exposed by the container application. Only applies to dev, on non-linux machines, and if the Dockerfile doesn't already declare exposed ports.
+	 */
+	dev_exposed_ports?: number[];
 };
 
 /**
@@ -717,6 +722,8 @@ export interface EnvironmentNonInheritable {
 		binding: string;
 		/** The name of the index. */
 		index_name: string;
+		/** Whether the Vectorize index should be remote or not (only available under `--x-mixed-mode`) */
+		remote?: boolean;
 	}[];
 
 	/**
@@ -799,6 +806,8 @@ export interface EnvironmentNonInheritable {
 	browser:
 		| {
 				binding: string;
+				/** Whether the Browser binding should be remote or not (only available under `--x-mixed-mode`) */
+				remote?: boolean;
 		  }
 		| undefined;
 
@@ -817,6 +826,8 @@ export interface EnvironmentNonInheritable {
 		| {
 				binding: string;
 				staging?: boolean;
+				/** Whether the AI binding should be remote or not (only available under `--x-mixed-mode`) */
+				remote?: boolean;
 		  }
 		| undefined;
 
@@ -834,6 +845,8 @@ export interface EnvironmentNonInheritable {
 	images:
 		| {
 				binding: string;
+				/** Whether the Images binding should be remote or not (only available under `--x-mixed-mode`) */
+				remote?: boolean;
 		  }
 		| undefined;
 
@@ -938,6 +951,8 @@ export interface EnvironmentNonInheritable {
 		namespace: string;
 		/** Details about the outbound Worker which will handle outbound requests from your namespace */
 		outbound?: DispatchNamespaceOutbound;
+		/** Whether the Dispatch Namespace should be remote or not (only available under `--x-mixed-mode`) */
+		remote?: boolean;
 	}[];
 
 	/**